backend/requirements.txt
Django==3.0.14
djangorestframework==3.11.2
djangorestframework-simplejwt==4.6
django-filter==2.1.0
jsonfield==3.1.0
psycopg2-binary>=2.9
django-cors-headers>=3.7
requests
Pillow>=8.0.0